requirements.txt
streamlit
pandas
supabase
fpdf2
python-dateutil
passlib
